modify: groupController 수정(줄 간격, Status Code) 및 주석 추가

+++ src/controllers/groupController.js
@@ -2,6 +2,7 @@ import { groupService } from "../services/groupService";
 import { addressToXY } from "../utils/addressToXY.js";
 
 const groupController = {
+  // 공구를 생성하는 함수
   createGroup: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -41,6 +42,7 @@ const groupController = {
     }
   },
 
+  // 공구의 수량 값을 변경하는 함수
   updateQuantity: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -48,8 +50,9 @@ const groupController = {
       const quantity = req.body.quantity;
 
       const checkState = await groupService.checkState({ groupId });
+      
       if (checkState === -1) {
-        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        const errorMessage = "모집 실패 (State: -1)";
         throw new Error(errorMessage);
       }
 
@@ -68,12 +71,13 @@ const groupController = {
         payload: updatedGroupInfo,
       };
 
-      res.status(200).json(body);
+      return res.status(201).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 특정 아이디의 지불 여부를 변경하는 함수
   updatePayment: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -82,8 +86,9 @@ const groupController = {
       const paymentMethod = req.body.paymentMethod;
 
       const checkState = await groupService.checkState({ groupId });
+
       if (checkState === -1) {
-        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        const errorMessage = "모집 실패 (State: -1)";
         throw new Error(errorMessage);
       }
 
@@ -103,25 +108,28 @@ const groupController = {
         payload: updatedGroupInfo,
       };
 
-      res.status(200).json(body);
+      return res.status(201).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 특정 아이디의 리뷰 여부를 변경하는 함수
   updateReview: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
       const review = req.body.review;
 
-      if (review === null) {
-        throw new Error("review 작성 여부(true/false)를 입력해주세요.");
+      if (!review) {
+        const errorMessage = "review 작성 여부(true/false)를 입력해주세요.";
+        throw new Error(errorMessage);
       }
 
       const checkState = await groupService.checkState({ groupId });
+
       if (checkState === -1) {
-        const errorMessage = "가뭄이 들은 당근밭입니다.";
+        const errorMessage = "모집 실패 (State: -1)";
         throw new Error(errorMessage);
       }
 
@@ -140,16 +148,16 @@ const groupController = {
         payload: updatedGroupInfo,
       };
 
-      res.status(200).json(body);
+      res.status(201).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // (공구 수량 값, 지불 여부, 리뷰 여부) 이외의 공동구매 정보 변경
   updateGroup: async (req, res, next) => {
     try {
       const groupId = req.params.groupId;
-
       const groupType = req.body.groupType ?? null;
       const groupName = req.body.groupName ?? null;
       const location = req.body.location ?? null;
@@ -157,8 +165,13 @@ const groupController = {
       const productId = req.body.productId ?? null;
       const state = req.body.state ?? null;
 
-      await groupService.checkState({ groupId });
+     const checkState = await groupService.checkState({ groupId });
 
+      if (checkState === -1) {
+        const errorMessage = "모집 실패 (State: -1)";
+        throw new Error(errorMessage);
+      }
+      
       const coordinates = await addressToXY(location);
       const toUpdate = {
         groupType,
@@ -184,12 +197,13 @@ const groupController = {
         payload: updatedGroup,
       };
 
-      res.status(200).json(body);
+      return res.status(201).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // (공동구매 아이디를 통해) 공동구매를 검색하는 함수
   getGroup: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -205,12 +219,13 @@ const groupController = {
         payload: groupInfo,
       };
 
-      res.status(200).json(body);
+      return res.status(200).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 상품에 대한 모든 공동구매들을 반환하는 함수
   getGroupsByProductId: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -228,16 +243,16 @@ const groupController = {
         payload: groupList,
       };
 
-      res.status(200).send(body);
+      return res.status(200).send(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 공동구매의 숫자 정보들을 반환하는 함수
   getGroupsByNumberInfo: async (req, res, next) => {
     try {
       const groupId = req.params.groupId;
-
       const numberInfo = await groupService.getNumberInfoByGroupId({ groupId });
 
       const body = {
@@ -245,12 +260,13 @@ const groupController = {
         payload: numberInfo,
       };
 
-      res.status(200).send(body);
+      return res.status(200).send(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 특정 조건(router 주석 참조)을 만족하는 공동구매를 반환하는 함수
   getGroupsByIsManager: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -265,23 +281,24 @@ const groupController = {
         payload: grouplist,
       };
 
-      res.status(200).send(body);
+      return res.status(200).send(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 공동구매에 참가자를 추가하는 함수
   updateParticipateIn: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
       const { quantity, paymentMethod } = req.body;
 
-      // const checkState = await groupService.checkState({ groupId });
-      // if (checkState === -1) {
-      //   const errorMessage = "가뭄이 들은 당근밭입니다.";
-      //   throw new Error(errorMessage);
-      // }
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "모집 실패 (State: -1)";
+        throw new Error(errorMessage);
+      }
 
       const UpdatedGroup = await groupService.addParticipants({
         userId,
@@ -298,22 +315,24 @@ const groupController = {
         success: true,
         payload: UpdatedGroup,
       };
-      res.status(200).json(body);
+
+      return res.status(201).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 공동구매에서 참가자를 내보내는 함수
   updateParticipateOut: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
 
-      // const checkState = await groupService.checkState({ groupId });
-      // if (checkState === -1) {
-      //   const errorMessage = "가뭄이 들은 당근밭입니다.";
-      //   throw new Error(errorMessage);
-      // }
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "모집 실패 (State: -1)";
+        throw new Error(errorMessage);
+      }
 
       const UpdatedGroup = await groupService.deleteParticipant({
         userId,
@@ -328,12 +347,14 @@ const groupController = {
         success: true,
         payload: UpdatedGroup,
       };
-      res.status(200).json(body);
+
+      return res.status(204).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 공동구매 상태값을 확인하는 함수
   getGroupStateInfo: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -349,12 +370,13 @@ const groupController = {
         payload: stateInfo,
       };
 
-      res.status(200).json(body);
+      return res.status(200).json(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 마감 기한이 24시간 이내인 공동구매를 오름차순으로 정렬하는 함수
   getGroupsByRemainedTime: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
@@ -367,35 +389,37 @@ const groupController = {
         payload: groupList,
       };
 
-      res.status(200).send(body);
+      return res.status(200).send(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 모집 인원이 3명 이하인 공동구매를 오름차순 정렬하는 함수
   getGroupsByRemainedPersonnel: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
-      const groupList =
-        await groupService.getSortedGroupsByRemainedPersonnelInfo(userId);
+      const groupList = await groupService.getSortedGroupsByRemainedPersonnelInfo(userId);
 
       const body = {
         success: true,
         payload: groupList,
       };
 
-      res.status(200).send(body);
+      return res.status(200).send(body);
     } catch (error) {
       next(error);
     }
   },
 
+  // 반경 5km 이내 공동구매 오름차순 정렬
   getGroupsByLocations: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
       const distanceOption = req.query.option ?? "5000";
       const groupList = await groupService.findNearGroupList({ userId, distanceOption });
-      if (groupList.data === false) {
+
+      if (!groupList.data) {
         const body = {
           success: true,
           data: groupList.data,
@@ -417,16 +441,17 @@ const groupController = {
     }
   },
 
+  // 공동구매를 논리 삭제하는 함수
   deleteGroup: async (req, res, next) => {
     try {
       const userId = req.currentUserId;
       const groupId = req.params.groupId;
 
-      // const checkState = await groupService.checkState({ groupId });
-      // if (checkState === -1) {
-      //   const errorMessage = "가뭄이 들은 당근밭입니다.";
-      //   throw new Error(errorMessage);
-      // }
+      const checkState = await groupService.checkState({ groupId });
+      if (checkState === -1) {
+        const errorMessage = "모집 실패 (State: -1)";
+        throw new Error(errorMessage);
+      }
 
       const UpdatedGroup = await groupService.deleteGroup({
         userId,
@@ -441,7 +466,8 @@ const groupController = {
         success: true,
         payload: UpdatedGroup,
       };
-      res.status(200).json(body);
+
+      return res.status(204).json(body);
     } catch (error) {
       next(error);
     }
